requirements.txt
textual>=0.41.0
pydantic>=2.0.0
pyyaml>=6.0
rich>=13.0.0
docker>=6.0.0
questionary>=1.10.0
psutil>=5.9.0
typer>=0.9.0
openai>=1.0.0
aiohttp>=3.8.0
websockets>=11.0.0
prometheus-client>=0.17.0
httpx>=0.24.0
click>=8.0.0
dataclasses-json>=0.5.0
asyncio-mqtt>=0.13.0
kubernetes>=27.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
redis>=4.6.0
